requirements.txt
amqp==2.1.4
ansible==2.2.0.0
apache-libcloud==1.3.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
boto==2.43.0
celery==4.0.2
cffi==1.8.3
cryptography==1.5.2
Django==1.10.3
django-celery-results==1.0.1
django-cors-headers==1.2.2
django-debug-toolbar==1.7
django-extensions==1.7.4
django-jsonfield==1.0.1
djangorestframework==3.5.4
django-rest-swagger==2.1.2
docker-py==1.10.5
docker-pycreds==0.2.1
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.20.0
flower==0.9.1
functools32==3.2.3-2
future==0.16.0
futures==3.0.5
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.18.3
google-api-python-client==1.5.4
google-apitools==0.5.5
google-compute-engine==2.2.4
google-gax==0.12.5
googleapis-common-protos==1.4.0
greenlet==0.4.11
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.0.1
gunicorn==19.6.0
httplib2==0.9.2
idna==2.1
ipaddress==1.0.17
Jinja2==2.8
jsonfield==1.0.1
jsonschema==2.5.1
kombu==4.0.2
MarkupSafe==0.23
meld3==1.0.2
MySQL-python==1.2.5
oauth2client==3.0.0
paramiko==2.0.2
ply==3.8
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
python-dateutil==2.5.3
pytz==2016.10
PyYAML==3.12
requests==2.11.1
rsa==3.4.2
setuptools==18.5
setuptools-git==1.1
six==1.10.0
SQLAlchemy==1.1.4
supervisor==3.3.1
uritemplate==3.0.0
vine==1.1.3
websocket-client==0.37.0
-e .
